chore: refactor the code and fix google refresh issue
