+++ .turbo/preferences/tui.json
@@ -1,4 +1,4 @@
 {
   "is_task_list_visible": true,
-  "active_task": null
+  "active_task": "one#dev"
 }
